some formate update

+++ app/Http/Controllers/Auth/TwoFactorController.php
@@ -13,6 +13,9 @@ class TwoFactorController extends Controller
 {
     public function varify(Request $request)
     {
+        if (!auth()->check()) {
+            return response()->json(['message' => 'register before'], 401);
+        }
         $user = auth()->user();
         $validator = Validator::make(request()->all(), [
             'code' => 'required'
@@ -25,29 +28,32 @@ class TwoFactorController extends Controller
         if (auth()->check()) {
             if ($user->expire_at < now()) {
                 $user->delete();
-                return response()->json(['timeout! sorry this code is not working please sign up again'], 400);
+                return response()->json(['message' => 'timeout! sorry this code is not working please sign up again'], 400);
             }
         }
 
         if ($request->code == $user->code) {
             $user->resetTwoFactorCode();
-            return response()->json(['thank you, you enter a correct code'], 200);
+            return response()->json(['message' => 'thank you, you enter a correct code'], 200);
         } else {
-            return response()->json(['sorry you enter uncorect code'], 400);
+            return response()->json(['message' => 'sorry you enter uncorect code'], 400);
         }
 
     }
     public function resendCode()
     {
+        if (!auth()->check()) {
+            return response()->json(['message' => 'register before'], 401);
+        }
         $user = auth()->user();
         if (auth()->check()) {
             if ($user->expire_at < now()) {
                 $user->delete();
-                return response()->json(['timeout! sorry this code is not working please sign up again']);
+                return response()->json(['message' => 'timeout! sorry this code is not working please sign up again']);
             }
         }
         $user->generateCode();
         Mail::to($user->email)->send(new TwoFactorMail($user->code, $user->name));
-        return response()->json(['your varification code resend to your email successfully'], 200);
+        return response()->json(['message' => 'your varification code resend to your email successfully'], 200);
     }
 }
